+++ app/src/main/java/space/jay/mvvm_with_clean_architecture/ui/activity/MainActivity.kt
@@ -7,8 +7,8 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.rememberNavController
 import dagger.hilt.android.AndroidEntryPoint
 import space.jay.mvvm_with_clean_architecture.core.common.window.Feature
-import space.jay.mvvm_with_clean_architecture._feature.pokemon.routeFinalPokemon
-import space.jay.mvvm_with_clean_architecture._feature.pokemon.toPokemon
+import space.jay.mvvm_with_clean_architecture.feature.pokemon.routeFinalPokemon
+import space.jay.mvvm_with_clean_architecture.feature.pokemon.toPokemon
 import space.jay.mvvm_with_clean_architecture.ui.theme.BaseTheme
 import java.lang.ref.WeakReference
