+++ feature/pokemon/src/main/AndroidManifest.xml
