+++ feature/searchWiki/src/main/AndroidManifest.xml
