+++ settings.gradle
@@ -20,5 +20,5 @@ include ":core:domain"
 include ":core:model"
 include ":core:network"
 include ":core:ui"
-include ':_feature:searchWiki'
-include ':_feature:pokemon'
+include ":feature:searchWiki"
+include ":feature:pokemon"
